+++ src/application/bill/bill.service.ts
@@ -515,33 +515,4 @@ export class BillService {
     if (bill) return bill
     else ErrorHandler.NOT_FOUND_MESSAGE('Bill not found')
   }
-
-  async getLastFiveBills(userId: string) {
-    try {
-      const thisDate = {
-        months: [new Date().getMonth()],
-        years: [new Date().getFullYear()]
-      }
-      const thisDates: Array<Date[]> = []
-      thisDate.years.forEach((y) =>
-        thisDate.months.forEach((m) => {
-          thisDates.push([new Date(y, m, 1), new Date(y, m + 1, 0)])
-        })
-      )
-      const bills = await this.billService.find({
-        relations: ['creditCard', 'company', 'bank1', 'bank2', 'category'],
-        where: {
-          due: Or(...thisDates.map((d) => Between(d[0], d[1]))),
-          userId
-        },
-        take: 5,
-        order: { updatedAt: 'DESC' }
-      })
-      return {
-        bills
-      }
-    } catch (error) {
-      return ErrorHandler.handle(error)
-    }
-  }
 }
